+++ CHANGES.md
@@ -1,3 +1,13 @@
+v0.5.0
+======
+
+* Support for cases similarity queries
+
+v0.4.0
+======
+
+* Paginated endpoints return a number when `count=True`
+
 v0.3.0
 ======